requirements.txt
# ── HTTP ────────────────────────────────────────
requests>=2.32

# ── Claude API ───────────────────────────────────
anthropic>=0.40

# ── 환경변수 ─────────────────────────────────────
python-dotenv>=1.0

# ── 국내 데이터 ──────────────────────────────────
FinanceDataReader>=0.9
pandas>=2.2
pandas-ta>=0.3

# ── 글로벌 데이터 ─────────────────────────────────
yfinance>=0.2
fredapi>=0.5

# ── 문서 빌드 (build_html.py) ────────────────────
markdown>=3.6
